# Conflicts:
#	src/main/java/ir/niopdc/policy/utils/GrpcUtils.java

+++ src/main/java/ir/niopdc/policy/domain/policy/PolicyService.java
@@ -4,5 +4,5 @@ import ir.niopdc.base.BaseService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PolicyService extends BaseService<PolicyRepository, Policy, Integer> {
+public class PolicyService extends BaseService<PolicyRepository, Policy, Byte> {
 }
